# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ran API
  description: Red team automation and attack campaign management API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /api/applicable-ttps:
    get:
      summary: Get applicable TTPs for a target
      description: Returns TTPs that can be applied to the specified target. If targetId is omitted or empty, returns TTPs for all targets.
      operationId: getApplicableTTPs
      parameters:
        - name: targetId
          in: query
          required: false
          schema:
            type: string
          description: ID of the target entity (optional, can be empty for all targets)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TTP'
        '404':
          description: Target not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/action/execute:
    post:
      summary: Execute action
      description: Execute a TTP action on a target. The action will be queued and executed asynchronously. The final execution status will be sent via SSE.
      operationId: executeAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteActionCmd'
      responses:
        '200':
          description: Action executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '202':
          description: Action queued for execution, pending asynchronous processing. The final status will be sent via SSE.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  taskId:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The action was executed, but failed for some reason
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/campaign/reset:
    post:
      summary: Reset campaign
      description: Reset the current campaign state
      operationId: resetCampaign
      responses:
        '200':
          description: Campaign reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '500':
          description: Reset failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pods/running:
    get:
      summary: Get running pods
      description: Returns list of running pods in the specified namespace
      operationId: getRunningPods
      parameters:
        - name: namespace
          in: query
          required: false
          schema:
            type: string
          description: Kubernetes namespace (optional, defaults to all namespaces)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/K8sResource'
        '500':
          description: Failed to retrieve pods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    TTP:
      type: object
      required:
        - id
        - name
        - description
        - tactic
        - techniques
        - requires
        - procedures
        - params
        - status
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        tactic:
          type: string
        techniques:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/TTPStatus'
        params:
          type: array
          items:
            $ref: '#/components/schemas/TTPParam'
        requires:
          $ref: '#/components/schemas/Requirements'
        effects:
          type: array
          items:
            type: string
        procedures:
          type: array
          items:
            $ref: '#/components/schemas/Procedure'
        defense:
          $ref: '#/components/schemas/TTPDefense'
    Procedure:
      type: object
      required:
        - id
        - command
      properties:
        id:
          type: string
        command:
          type: string
        tool:
          type: string
        isLocalCommand:
          type: boolean
    TTPParam:
      type: object
      required:
        - name
        - type
        - required
        - description
        - default
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        required:
          type: boolean
        default:
          type: string
    TTPDefense:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        description: 
          type: string
        sigma:
          $ref: '#/components/schemas/SigmaRule'
        d3efend:
          type: string
    SigmaRule:
      type: object
      required:
        - title
        - tags
        - logsource
        - detection
        - date
      properties:
        title:
          type: string
        id:
          type: string
        description:
          type: string  
        status:
          type: string
        logsource:
          type: object
          additionalProperties: true
        detection:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
        references:
          type: array
          items:
            type: string
        date:
          type: string
          format: date-time
    ExecuteActionCmd:
      type: object
      required:
        - actionId
        - targetId
      properties:
        actionId:
          type: string
        targetId:
          type: string
        procedureId:
          type: string
        args:
          type: object
          additionalProperties:
            type: string
    K8sResource:
      type: object
      required:
        - id
        - name
        - kind
      properties:
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        kind:
          type: string
    Requirements:
      type: object
      properties:
        kind:
          type: string
        rbacPermissions:
          type: array
          items:
            $ref: '#/components/schemas/RBACPermission'
        accessLevel:
          type: string
        exists:
          type: array
          items:
            type: string
    RBACPermission:
      type: object
      required:
        - apiGroup
        - verb
        - resource
        - resourceName
        - resourceType
        - scope
        - sourceRole
      properties:
        resource:
          type: string
        resourceName:
          type: string
        resourceType:
          type: string
        apiGroup:
          type: string
        verb:
          type: string
        scope:
          type: string
        sourceRole:
          type: string
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          type: string
    TTPStatus:
      type: string
      enum:
        - enabled
        - disabled
      description: Status of a TTP - whether it is enabled or disabled